+++ requirements/models.txt
@@ -1,3 +1,4 @@
 torchvision>=0.7
 scikit-learn>=0.23
+Pillow
 opencv-python
